+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Nothing here yet.
 - Time specification "nicknames" support, e.g. `@yearly`, `@monthly`,
   `@daily`, etc. ([32](https://github.com/pilosus/kairos/issues/32))
 
+- Explicit testing for GraalVM native-image compilation
+  ([38](https://github.com/pilosus/kairos/issues/38))
+
 ## [v0.2.45] - 2025-12-16
 
 ### Added